Fix a bug where the correlator would crash when two consecutive calls to initialize_next_scan were made (without calling create_akima_spline).
This can happen when correlating a schedule with very short scans. In this 
case  it is possible that a correlator node receives a UVW table for a 
certain scan but not compute a single correlation from that scan. The
create_akima_spline method is called when a UVW value is computed.


git-svn-id: svn+ssh://code.jive.nl/code/svn/sfxc/trunk@1939 5736e3ea-d426-0410-8aaf-93e336ac3372