Replace blocking MPI calls with non-blocking calls on the input node.
Blocking MPI calls use a spinlock to wait for messages, 
in theory this shouldn't affect performance because it 
is supposed to call sched_yield(). 

Unfortunatly, the spinlock doesn't seem to be performing
as advertised. For a certain 2gbps diskbased experiment
we got a 100% increase in speed after replacing the 
blocking MPI calls. Furthermore, without the spinlock
the cpu usage is much easier to profile.


git-svn-id: svn+ssh://code.jive.nl/code/svn/sfxc/trunk@1943 5736e3ea-d426-0410-8aaf-93e336ac3372