For mixed bandwidth baselines the weights were incorrect when there was invalid data. Because we use the weights to scale the amplitudes this also means that the amplitudes were wrong. But only if there is invalid data.
For every baseline (station1, station2) the method find_invalid determines
if ranges of invalid data for station1 overlap with invalid ranges
from station2. We use the sample number within the integration for this.
But we didn't take into account that the sample rates of station1 and 
station2 can potentially be different.



git-svn-id: svn+ssh://code.jive.nl/code/svn/sfxc/trunk@2009 5736e3ea-d426-0410-8aaf-93e336ac3372