In vex2 there will no longer be a sample_rate field in the $FREQ section, instead this information is obtained from $TRACKS / $BITSTREAMS / $DATASTREAMS.
This requires the correlator to know what data format a station is using which
the correlator obtains from the $DAS section. However, when using a virtual
setup station (e.g. for mixed bandwith or zoom bands) we don't add it to
$STATION to not complicate the j2ms2 step, which there is no $DAS for
the reference station. This crashes the correlator.

As a temporary work around, for vex 1.5 files we return the sample_rate 
from the $FREQ section without looking at $DAS.