Reverted back to using gnpuplot for the ftp fringeplots.
Giza, the toolkit we used previously, was creating way to large filesizes. 
This is because CAIRO can only output 24 bit PNGs, but for plotting you really
only need 8 bit PNGs. Of course we could have moved to plain PGPLOT as well
but this produces very bad looking plots due to a lack of anti-aliassing.



git-svn-id: svn+ssh://code.jive.nl/code/svn/sfxc/trunk@2243 5736e3ea-d426-0410-8aaf-93e336ac3372