Fix a deadlock that occurred in spectral line eVLBI.
When correlating with a large number of spectral points a single integration 
will not fit into an output node buffer, so when integration_write is
called the correlator will block on this write. Normally, this is purely a 
performance concern. The correlator node will unblock when the output node
starts writing the integration to disk.

However, there is a second route at which correlator nodes can enter a blocked
state. When a new scan is started the delay tables are distributed
using MPI_Bcast. This MPI routine uses a tree algorithm to distribute
the data. All nodes which are on the same branch of this tree algorithm
will block until all nodes on that branch join the MPI_Bcast.

Suppose now that the correlator node responsible for an integration at time
t1 has finished correlating and is blocking in integration_write, but
an earlier integration at some t0 (i.e. t0 < t1) is still correlating. If at
this time a new scan is started *and* both nodes are on the same branch
of the MPI_Bcast tree then the node responsible for t0 will now also start
blocking. Because the t0 node is waiting for the t1 node to join the Bcast, 
but the t1 node is in turn waiting for the t0 to finish correlating a
deadlock occurs.

The solution for this issue is to refactor the code so the MPI calls are
handled in a different thread than the main correlator loop. To this end
the Correlator_node_tasklet class was added which executes the main correlator
loop, while all mpi calls are handles in the Correlator_node class.



git-svn-id: svn+ssh://code.jive.nl/code/svn/sfxc/trunk@2246 5736e3ea-d426-0410-8aaf-93e336ac3372