Fix crashes that occured for large FFTs in combination with sliced integrations.
In sliced integrations not every slice has an equal number of FFTs. There was a
race condition which caused the bit2float conversion to use the number of ffts per
slice from the previous integration.

Whenever a new slice is available all new parameters, are copied to the bit2float worker
and a boolean have_new_parameters is set to indicate that a new slice is available. In the
code the boolean is set after copying the data, so no race condition would exist. However,
the compiler reorders instructions so that the boolean is actually set before all parameters
are copied over.

To solve this issue the new parameters are now placed in a threadsafe queue.