Fix issue with the one letter station name mapping.
When a station is not in the one letter code file the following heuristic is used:
  1. Preferentially a code is used which is not in the one letter code file
  2. When we run out of codes in 1. then we re-use codes which are in the one 
     letter code file but are not present in the VEX file.

The code to find the stations that are not in the VEX file was broken, it didn't 
handle the case where mulitple stations are linked to the same one letter code. For
example in the default mapping Ef and Ed get the same one letter code.