Add option "output_buffer_size" which allows to specify the size in bytes of the per correlator-node buffers on the output node.
For optimal performance the buffer should be large enough to hold at 
least one full integration. If this is not the case correlator-nodes may
block for long periods of time while they wait until it is their turn
to write data. 

The default value is 1250000 bytes, which is the size of the buffers in 
previous versions of SFXC.