Reading legacy VDIF was broken.
We determine if a station is using legacy VDIF by inspecting the
first frame. However, this causes problems if the first frame is 
invalid. In these cases SFXC doesn't know the size of the header,
and will assume normal (non-legacy) VDIF.

We read VDIF frames as blocks of multiple frames for efficiency.
So when reading a new block we check if the first frame coincides with
the first frame of a block, otherwise we reject it.

What was going wrong was that if the first frame was valid, but not
at the start of a new block we would still mark the frame as invalid. 
When this happens to the first frame of a legacy VDIF data stream this
would cause the correlation to fail.


git-svn-id: svn+ssh://code.jive.nl/code/svn/sfxc/trunk@2325 5736e3ea-d426-0410-8aaf-93e336ac3372