gfortran 10+ now does more stringent type checking which causes compilation of CALC to fail.
Some functions such as e.g. put4 take an array as argument but the length of the array depends on
the item being pushed. When there is only one value then put4 may be called with a scalar,
which is then interpreted by put4 as an array of length 1. This conversion between scalars
and arrays of length one works because fortran passes parameters by reference.

Starting gfortran version 10 this behaviour it is no longer allowed by default, though an error
is only triggered when you call a function both with an array and a scalar in the same routine.
In this case gfortran will trow a "Rank mismatch" error. 

As a quick fix add the -fallow-argument-mismatch flag to FFLAGS when a newer gfortran is detected. 
This flag converts the rank mismatch errors into warnings, a more permanent fix would be to convert 
the scalars involved to arrays of length 1.


git-svn-id: svn+ssh://code.jive.nl/code/svn/sfxc/trunk@2341 5736e3ea-d426-0410-8aaf-93e336ac3372