Also commit source name length fix to bolometer. ------------------------------------------------------------------------ r2336 | keimpema | 2023-02-01 14:14:12 +0100 (Wed, 01 Feb 2023) | 7 lines
In Correlation_parameters we had a source name size limit of 10 characters,
however, VEX specifies a maximum of 16 characters.  This size limit is
increaded to 16 characters.

The source name size limits were correct in the rest of the code so this
change doesn't affect the source name lengths in the global headers.